GamePlayFragment.kt
RollDice 객체를 생성하여
각 플레이어가 주사위를 굴리고, turn을 주고받는 형식의 게임을 만들 수 있도록 한다.

RollDice의 프로퍼티 'turn'은 플레이어가 3번 주사위를 굴리고 나면 !turn으로 값을 갱신한다.

first Player는 turn==true && player == firstPlayer일 때 자신의 turn이 돌아오고,
second Player는 turn==false && player == secondPlayer일 때 자신의 turn이 돌아온다.

turn이 돌아올 때 reloadRollDice() 메서드를 실행하여 diceList, keepList, chance (각각 주사위 목록, 고정한 주사위의 목록, 굴릴 수 있는 횟수를 의미) 변수값을 모두 refresh한다.

chance가 끝나면 (0==chance) 해당 플레이어는 rollDice()에 turn을 갱신하여 firebase database에 write -> 다음 플레이어 시작.

+++ app/src/main/java/com/ranseo/yatchgame/data/model/Board.kt
@@ -39,21 +39,21 @@ data class Board(
 ) {
 
     constructor(hashMap:HashMap<*,*>): this(
-        hashMap["ones"] as Int,
-        hashMap["twos"] as Int,
-        hashMap["threes"] as Int,
-        hashMap["fours"] as Int,
-        hashMap["fives"] as Int,
-        hashMap["sixes"] as Int,
-        hashMap["sum"] as Int,
-        hashMap["bonus"] as Int,
-        hashMap["choice"] as Int,
-        hashMap["fourCard"] as Int,
-        hashMap["fullHouse"] as Int,
-        hashMap["smallStraight"] as Int,
-        hashMap["largeStraight"] as Int,
-        hashMap["yacht"] as Int,
-        hashMap["total"] as Int
+        (hashMap["ones"] as Long).toInt(),
+        (hashMap["twos"] as Long).toInt(),
+        (hashMap["threes"] as Long).toInt(),
+        (hashMap["fours"] as Long).toInt(),
+        (hashMap["fives"] as Long).toInt(),
+        (hashMap["sixes"] as Long).toInt(),
+        (hashMap["sum"] as Long).toInt(),
+        (hashMap["bonus"] as Long).toInt(),
+        (hashMap["choice"] as Long).toInt(),
+        (hashMap["fourCard"] as Long).toInt(),
+        (hashMap["fullHouse"] as Long).toInt(),
+        (hashMap["smallStraight"] as Long).toInt(),
+        (hashMap["largeStraight"] as Long).toInt(),
+        (hashMap["yacht"] as Long).toInt(),
+        (hashMap["total"] as Long).toInt()
     )
